requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.16
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
exifread==3.0.0
gunicorn==20.1.0
idna==3.4
image==1.5.33
Markdown==3.3.4
oauthlib==3.2.2
Pillow==9.2.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
httpie==3.2.1
imutils==0.5.4
opencv-python